Move into one vite project
