+++ example/main.tsx
@@ -4,6 +4,8 @@ import React from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App";
 
-const container = document.getElementById("root");
-const root = createRoot(container!);
-root.render(<App />);
+createRoot(document.getElementById("root")!).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
