+++ index.html
@@ -21,6 +21,6 @@
   <body>
     <noscript> You need to enable JavaScript to run this app. </noscript>
     <div id="root"></div>
-    <script type="module" src="/src/index.tsx"></script>
+    <script type="module" src="/example/main.tsx"></script>
   </body>
 </html>
