+++ package.json
@@ -33,8 +33,8 @@
     "task"
   ],
   "scripts": {
-    "build": "tsc && vite build",
-    "start": "microbundle watch --jsx React.createElement --tsconfig ./tsconfig.json --compress",
+    "dev": "vite",
+    "build": "tsc --p ./tsconfig-build.json && vite build",
     "test:lint": "eslint .",
     "test:prettier": "prettier --check ./src",
     "test:unit": "vitest run",
