+++ tsconfig.json
@@ -22,8 +22,9 @@
     "forceConsistentCasingInFileNames": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true
+    "noEmit": true,
+    "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "example", "src/test"]
+  "include": ["example", "src"],
+  "references": [{ "path": "./tsconfig.node.json" }]
 }
